database synchronization bugfix + included playlists activity
Database was always up to date because of wrong stored values.
App will support any kind of playlist, this is the start.

+++ android_app/src/de/viktorreiser/bansheeremote/data/BansheeConnection.java
@@ -406,8 +406,8 @@ public class BansheeConnection {
 				return params[0] == 2;
 			}
 			
-			public static long decodeFileTimestamp(byte [] response) {
-				return response.length < 4 ? 0 : decodeInt(response, 0);
+			public static int decodeFileTimestamp(byte [] response) {
+				return response.length < 4 ? 0 : (int) decodeInt(response, 0);
 			}
 		}
